+++ src/app/convenience-file-types.js
@@ -52,6 +52,25 @@ export function expandClustersJson(fileDef) {
     {
       ...baseFileDef,
       fileType: FileType.OBS_FEATURE_MATRIX_CLUSTERS_JSON,
+      fileType: FileType.OBS_FEATURE_MATRIX_GENES_JSON,
+    },
+  ];
+}
+
+export function expandGenesJson(fileDef) {
+  const baseFileDef = {
+    ...fileDef,
+    coordinationValues: {
+      ...fileDef.coordinationValues,
+      obsType: fileDef.coordinationValues?.obsType || 'cell',
+      featureType: fileDef.coordinationValues?.featureType || 'gene',
+    },
+  };
+  delete baseFileDef.type;
+  return [
+    {
+      ...baseFileDef,
+      fileType: FileType.OBS_FEATURE_MATRIX_GENES_JSON,
     },
   ];
 }
@@ -62,5 +81,6 @@ export function expandClustersJson(fileDef) {
  */
 export const CONVENIENCE_FILE_TYPES = {
   // [FileType.CLUSTERS_JSON]: expandClustersJson,
+  // [FileType.GENES_JSON]: expandGenesJson,
   // [FileType.CELLS_JSON]: expandCellsJson,
 };