+++ src/app/convenience-file-types.test.js
@@ -1,5 +1,6 @@
 import expect from 'expect';
 import {
+  expandGenesJson,
   expandCellsJson,
   expandClustersJson,
 } from './convenience-file-types';
@@ -21,6 +22,24 @@ describe('src/app/convenience-file-types.js', () => {
         },
       ]);
     });
+  });
+  describe('expandGenesJson', () => {
+    it('expands', () => {
+      expect(expandGenesJson({
+        fileType: 'genes.json',
+        url: 'http://localhost:8000/genes.json',
+      })).toEqual([
+        {
+          fileType: 'obsFeatureMatrix.genes.json',
+          url: 'http://localhost:8000/genes.json',
+          coordinationValues: {
+            obsType: 'cell',
+            featureType: 'gene',
+          },
+        },
+      ]);
+    });
+  });
   describe('expandCellsJson', () => {
     it('expands when there are no options', () => {
       expect(expandCellsJson({