+++ src/loaders/index.js
@@ -1,7 +1,7 @@
 export { default as JsonLoader } from './JsonLoader';
 export { default as MatrixZarrLoader } from './MatrixZarrLoader';
-export { default as GenesJsonAsMatrixZarrLoader } from './GenesJsonAsMatrixZarrLoader';
 export { default as ClustersJsonAsMatrixZarrLoader } from './clusters-json-loaders/ClustersJsonAsMatrixZarrLoader';
+export { default as GenesJsonAsMatrixZarrLoader } from './genes-json-loaders/GenesJsonAsMatrixZarrLoader';
 export { default as AnnDataLoaders } from './anndata-loaders';
 export { default as LoaderResult } from './LoaderResult';
 export { default as AbstractTwoStepLoader } from './AbstractTwoStepLoader';