+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Added `npm run start:nolint` script to disable linting for quickly prototyping code.
 - Added new file type and data type constants.
 - Added loaders for `obsFeatureMatrix.clusters.json`
+- Added loaders for `obsFeatureMatrix.genes.json`
 - Added loaders for `obsLabels.cells.json`, `obsEmbedding.cells.json`, `obsLocations.cells.json`, `obsSegmentations.cells.json`
 
 ### Changed